[dtls] handle events when disconnected (#115)
This commit fixes DTLS events handling when the session is
disconnected. Previous implementation has the assumption that the
underlying socket will be closed when the session is disconnected. But
now, the socket now may be open in this case.

This commit ignores the incoming data when the DTLS session is
disconnected.