[active-dataset] returns raw Active Operational Dataset TLVs (#142)
Parsing the Active Operational Dataset into struct
ActiveOperationalDataset is for manipulating each field of the dataset
and the commissioner app doesn't need to write a decoder by their
own. But the commissioner app may store or transfer Active Operational
Dataset as a byte array. It would be convenient to have the raw TLV
format of the dataset so that the commissioner app doesn't need to
write an encoder by their own.

This commit adds a new APIs (GetRawActiveDataset(...)) to return raw
Active Operational Dataset.