[swig] map Java byte to uin8_t for only the ByteArray class (#151)
This commit helps SWIG to map Java byte to/from C++ uint8_t when it is
used for std::vector<uint8_t>.

SWIG by default maps Java short to/from C++ uint8_t because Java byte
is a signed integer that cannot hold all values of uint8_t. But when
it comes to a byte array, it doesn't matter whether it is signed or
unsigned and we'd like to map uint8_t to/from Java byte because they
have the same width.