[swig] revert changes to the minimum SWIG version (#153)
This commit reverts changes of downgrading minimum SWIG version to 3.0.

We downgraded the minimum version for easy installation of SWIG on old
Linux distributions. But the incoming Commissioner Android App relies
on the Java methods generated by SWIG 4.0. Unfortunately, we didn't
have a GH job to test the SWIG-generated Java code for now, but it
will come with the Android App.