[dataset] avoid setting session ID or Border Agent Locator (#212)
Commissioner Session ID and Border Agent Locator in the Commissioner
Dataset are read-only.  Current implementation rejects setting the
Commissioner Dataset if any of the fields is present, but it is not
convenient for the caller to manually clear out those fields. This
commit change the implementation to always clear out the Commissioner
Session ID and Border Agent Locator fields before sending the
MGMT_COMM_SET.req message to the border agent.