[cli] remove the macro definition of `SO_BINDTODEVICE` from interpreter module
A previously introduced macro `SO_BINDOTDEVICE`, intended to pass the
pre-commit code checks, which is conflicted with an definition in header
file named <sys/socket.h>. So we need remove the macro from the
interpreter module. To ensure the pre-commit code checks pass, we need
to include the approriate header file and call `setsocketopt` with
OS-specific parameters.