[cli] remove definition of `SO_BINDTODEVICE` from interpreter module (#263)
A previously introduced macro `SO_BINDOTDEVICE`, intended to pass the
pre-commit code checks, which is conflicted with an definition in
header file named <sys/socket.h>. So we need remove the macro from the
interpreter module. To ensure the pre-commit code checks pass, we need
to include the appropriate header file and call `setsocketopt` with
OS-specific parameters.