[api] enable DIAG_GET query/answer commands into commissioner
The DIAG_GET query and answer commands are used for diagnostic TMF
messages transaction, which can support to query the child tabler and
router neighbor from the queried devices for generating the network
topology.

In this transaction, the inquirer will send query message to the peer
device with specified TLV IDs, the peer device will respond answer
message to the inquirer device with TLVs data.