[udp-proxy] set local port to dedicated port number(:mm) for udp proxy transaction (#279)
Currently, while the commissioner is transmitting udp_tx messages, the
UDP source port is provided by the `DtlsSession` interface. This value
is randomly generated. However, the CoAP layer on the border router
side performs checks on the received udp_tx port to ensure that the
destination UDP port used to outgoing messages matches the source port
of the received messages. This leads to the commissioner's ACK
messages not being processed, causing the BR side to repeatedly
retransmit the same udp_rx message.

To resolve this issue, we use the default MM port as the only
supported destination port for UDP_RX messages.