[android] add basic tests for auto generated java lib (#290)
When the SWIG interface file (commissioner.i) is updated, the generated
java library API may be broken. There are currently no tests to guard
this. This commit adds a basic androidTest for this purpose. Note that
an Android rather than junit test is added because the test case depends
on an JNI library.