[diag] add prefix tlv definition and parsing into diagnostic data
The Prefix TLV is a TLV included in Network data. This TLV records
information about alll border routers in the network and route
informaiton to external networks. It is helpful for users to understand
the routing status between the Thread network and external networks.

Prefix TLV has three types of sub-tlv:

* hasRoute TLV
* BorderRouter TLV
* 6LoWPAN context ID

Test: unit test