[python] explicitly specify python3 as the python interpreter (#315)
There are places in the repo that uses `python <script>` command to
execute the code. This can cause problems when there's python2
interpreter available on the device because the code is for python3
only. This already happens on the OS of ot-reference-release because
the image is too old and python2 is still available on that one.

While there's still a distance before ot-reference-release can migrate
to a newer image, we need to fix it here.
There are places in the repo that uses `python <script>` command to
execute the code. This can cause problems when there's python2
interpreter available on the device because the code is for python3
only. This already happens on the OS of ot-reference-release because
the image is too old and python2 is still available on that one.

While there's still a distance before ot-reference-release can migrate
to a newer image, we need to fix it here.