[github-action] handl cmake installation confilict
In current GitHub actions, the runner will pre-install cmake from custom
repository. In this change, we need to uninstall the pre-installed cmake
to avoid the following error in bootstrap phase.

```
Error: cmake was installed from the local/pinned tap
but you are trying to install it from the homebrew/core tap.
```